Move BinPath and populate_venv_with_pex to `pex.venv`.
